requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
djangorestframework==3.16.1
idna==3.11
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
